LexemeHandlerTest: use inheritable EntityTypeDefinitions
EntityHandlerTestCase exposes its EntityTypeDefinitions configuration
now to allow extensions to add to it without reloading the configuration
files and not having to care about its instantiation itself.

Previously this test did not register core wikibase's EntityTypeDefinitions
(e. g. items) and this worked because during the instantiation
of the WikibaseRepo singleton, these are automatically added. With the
upcoming switch to services (see T265984) we gain better control and
can/have to explicitly define the desired mock data without relying on
side effects.

Depends-On: Ifa9b7064cf83ef3908dd98737071a45d49173aa6
Bug: T265984
Change-Id: Icc61135b3a68c03cac9bdcec818efd008cc326ab