Replace LanguageFallbackLabelDescriptionLookupFactory
FallbackLabelDescriptionLookupFactory produces a lookup that also
resolves redirects and has some caching built-in.

In particular, this will resolve redirects in wbsearchentities search
result descriptions of Lexemes and Forms if any of the items used for
Language, Lexical Category or Feature is a redirect.

This also now directly depends on the trait LexemeDescriptionTestCase
from Lexeme in order to not duplicate its logic between the two
repositories.

To ensure there is no silent breakage due to that, this patch also
depends on a CI infrastructure patch that makes the
WikibaseLexemeCirrusSearch tests always run on any WikibaseLexeme patch.

Bug: T312224
Depends-On: Id623328660fc5b12db46fa5ba67f0e3712ac223b
Depends-On: I7b3c60098b801f31fbcc46f0a8bc7355117a0708
Change-Id: I1bc87f7f5ad204f0156bc777cc349f8e91fd8f3d