Resolve redirects in LexemeDescription for Special:Search
FallbackLabelDescriptionLookupFactory produces a lookup that also
resolves redirects and has some caching built-in.

In particular, this will resolve redirects in Special:Search search
result descriptions of Lexemes if any of the items used for
Language or Lexical Category is a redirect.

This now also uses the trait for helping with testing LexemeDescription
directly from WikibaseLexeme. Since the local copy of that trait is now
unused, it is removed in this change.

Bug: T312224
Change-Id: I13eb7c74ef52229c71d6aaf4b95547a906e30b38