Rank lemma's over forms
Entity search uses constant score queries for different levels of
matching the lexeme, and uses popularity metrics to rank within
the groups created by the constant score queries. Adding a tie-breaker
allows to also take into account how many different fields are
matching.

Aims to fix an issue where low-popularity lexemes that match both the
lemma and a representation are ranked at the very bottom, below
lexemes that only match on representation.

The value of 0.02 was chosen mostly arbitrarily. It should be low
enough to prevent lower groups from pushing into the upper groups,
and seems to be large enough to fix a few example queries. It could
have negative effects on other use cases.

Additionally changes the default form-discount from 1 (no discount)
to 0.95 which provides a small discount. As we have a single lemma
and multiple forms, it seems reasonable to weight the lemme a little
more as users will probably expect to be able to directly find it
by this value.

Bug: T348877
Change-Id: I128d5d8c86dd907ae2878c245c4c33168fbb8661