WikibaseRepoEntityTypesHookHandler: Don’t reorder entity types
We need to use the search entity type definitions as the base of the
addition, and add the original entity type definitions onto it (not vice
versa), so that the search callback from the search entity type
definitions will override the original one. However, the way this was
previously implemented also meant that the resulting array would start
with the keys from the search entity type definitions, and only then any
other keys from the original definitions, potentially changing the order
of the array. This side effect is both undesirable and visible to users
in some cases, such as in the order the entity types are listed in the
scoped typeahead search. Rewrite the code to do the array addition in a
different way that preserves the order of the original array, and then
add some tests for the hook handler as well.

The specific effect of this hook handler was that the WikibaseLexeme
entity types – lexeme, form, sense – would move to the beginning of the
entity type definitions, before items and properties. (Of these three,
only lexeme was actually visible in the scoped typeahead search.)

See also WikibaseCirrusSearch change I206534998d.

Bug: T398817
Change-Id: If2277f42cc367fe491c073b6614e47f295293f87